refactor: divided ui elements into folders
