+++ example/lib/presentation/ui/showcase/all_widgets_stand.dart
@@ -4,6 +4,8 @@ import 'package:example/presentation/ui/showcase/input_fields.dart';
 import 'package:example/presentation/ui/showcase/input_fields_list.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:shuffle_uikit/ui_kit/content_wrappers/ui_kit_border_wrapper.dart';
+import 'package:shuffle_uikit/ui_kit/organisms/organisms.dart';
 import 'package:shuffle_uikit/shuffle_uikit.dart';
 import 'package:lottie/lottie.dart';
 
@@ -129,7 +131,16 @@ class _AllWidgetsStandState extends State<AllWidgetsStand> {
             mainAxisAlignment: MainAxisAlignment.start,
             children: [
               SpacingFoundation.verticalSpace16,
-              const UiKitFingerPrintSlider(),
+              FingerPrintSlider(
+                child: FingerPrintButton(
+                  height: 0.27.sw * 1.68 - 8.h,
+                  path: 'assets/animation_touch_id.json',
+                  title: Text(
+                    'Guess',
+                    style: context.uiKitTheme?.boldTextTheme.subHeadline,
+                  ),
+                ),
+              ),
               SpacingFoundation.verticalSpace16,
               UiKitLeadingRadioTile(
                 selected: selection,
@@ -1423,10 +1434,13 @@ class _AllWidgetsStandState extends State<AllWidgetsStand> {
   }
 }
 
-class UiKitFingerPrintSlider extends StatelessWidget {
-  const UiKitFingerPrintSlider({
+class FingerPrintSlider extends StatelessWidget {
+  const FingerPrintSlider({
     super.key,
+    required this.child,
   });
+
+  final Widget child;
 
   @override
   Widget build(BuildContext context) {
@@ -1435,7 +1449,7 @@ class UiKitFingerPrintSlider extends StatelessWidget {
 
     return Stack(
       children: [
-        CardWithBorderWrapper(
+        UiKitBorderWrapper(
           height: height,
           child: ClipRRect(
             borderRadius: BorderRadiusFoundation.all28,
@@ -1451,164 +1465,8 @@ class UiKitFingerPrintSlider extends StatelessWidget {
             ),
           ),
         ),
-        UiKitFingerPrintButton(
-          height: height - 8.h,
-          path: 'assets/animation_touch_id.json',
-          title: Text(
-            'Guess',
-            style: context.uiKitTheme?.boldTextTheme.subHeadline,
-          ),
-        ),
+        child,
       ],
-    );
-  }
-}
-
-class UiKitFingerPrintButton extends StatefulWidget {
-  const UiKitFingerPrintButton({
-    super.key,
-    required this.height,
-    required this.path,
-    required this.title,
-    this.width,
-    this.onPressed,
-  });
-
-  final String path;
-  final Widget title;
-  final double height;
-  final double? width;
-  final VoidCallback? onPressed;
-  @override
-  State<UiKitFingerPrintButton> createState() => _UiKitFingerPrintButtonState();
-}
-
-class _UiKitFingerPrintButtonState extends State<UiKitFingerPrintButton>
-    with TickerProviderStateMixin {
-  bool _isPressed = false;
-  late final AnimationController _controller;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = AnimationController(
-      duration: const Duration(seconds: 2),
-      vsync: this,
-    );
-    _setAnimationListener();
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  void _setAnimationListener() {
-    _controller.addStatusListener((status) {
-      if (status == AnimationStatus.completed) {
-        setState(() {
-          _isPressed = true;
-          widget.onPressed;
-        });
-      } else if (status == AnimationStatus.dismissed) {
-        setState(() {
-          _isPressed = false;
-        });
-      }
-    });
-  }
-
-  void _startAnimation() {
-    _controller.forward();
-  }
-
-  void _reverseAnimation() {
-    _controller.reverse();
-  }
-
-  List<BoxShadow>? _getShadow(bool isPressed) {
-    if (_isPressed) {
-      return const [
-        BoxShadow(
-          color: ColorsFoundation.shadowPink,
-          blurRadius: 10,
-          spreadRadius: -5,
-          offset: Offset.zero,
-        ),
-      ];
-    } else {
-      return [];
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTapDown: (_) => _startAnimation(),
-      onTapUp: (_) => _reverseAnimation(),
-      child: AnimatedContainer(
-        duration: const Duration(seconds: 0),
-        decoration: BoxDecoration(
-          borderRadius: BorderRadiusFoundation.all28,
-          boxShadow: _getShadow(_isPressed),
-        ),
-        child: UiKitCardWrapper(
-          width: widget.width ?? 106.w,
-          height: widget.height,
-          color: context.uiKitTheme?.cardColor ?? ColorsFoundation.surface3,
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              GradientableWidget(
-                gradient: GradientFoundation.touchIdLinearGradient,
-                child: widget.title,
-              ),
-              SpacingFoundation.verticalSpace12,
-              ClipRRect(
-                borderRadius: BorderRadius.circular(50),
-                child: SizedBox(
-                  height: 48.w,
-                  width: 48.w,
-                  child: FittedBox(
-                    fit: BoxFit.cover,
-                    child: LottieBuilder.asset(
-                      widget.path,
-                      fit: BoxFit.cover,
-                      controller: _controller,
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ).paddingAll(EdgeInsetsFoundation.all4),
-      ),
-    );
-  }
-}
-
-class CardWithBorderWrapper extends StatelessWidget {
-  const CardWithBorderWrapper({
-    super.key,
-    required this.child,
-    required this.height,
-  });
-
-  final double height;
-  final Widget child;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: height,
-      clipBehavior: Clip.hardEdge,
-      decoration: BoxDecoration(
-        borderRadius: BorderRadiusFoundation.all28,
-        border: GradientFoundation.touchIdgradientBorder,
-        color: context.uiKitTheme?.cardTheme.color,
-      ),
-      child: child,
     );
   }
 }
