+++ lib/ui_kit/organisms/organisms.dart
@@ -1,6 +1,7 @@
 /// This file exports all the organisms in the UI Kit.
 
 export '../molecules/place/base_ui_kit_media_widget.dart';
+export 'animated_buttons/finger_print_button.dart';
 export 'animated_buttons/bouncing_blur_button.dart';
 export 'animated_buttons/small_bouncing_blur_icon_button.dart';
 export 'bottom_navigation_bar/gradient_bottom_navigation_bar.dart';
